requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dictdiffer==0.9.0
docker==7.0.0
ecdsa==0.19.0
Faker==24.11.0
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
iso8601==1.1.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pypika-tortoise==0.1.6
pytest==8.1.1
pytest-alembic==0.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
text-unidecode==1.3
tomlkit==0.12.5
tortoise-orm==0.21.2
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
